[Utils] Integrata utility 'sonarqube'. Effettuata prima analisi sul progetto tools/utils

+++ tools/utils/src/org/openspcoop2/utils/id/apache/serial/TimeBasedAlphanumericIdentifierGenerator.java
@@ -137,7 +137,7 @@ public class TimeBasedAlphanumericIdentifierGenerator extends AbstractStringIden
 
     @Override
 	public long maxLength() {
-        return MAX_LONG_ALPHANUMERIC_VALUE_LENGTH + this.postfixSize;
+        return ((long)MAX_LONG_ALPHANUMERIC_VALUE_LENGTH) + ((long)this.postfixSize);
     }
 
     @Override
